Refactor -
Rename the two installPackage functions to be more descriptive

Move package installer types and creation of installer map into ./installers directory
