+++ src/installers/trpc.ts
@@ -2,14 +2,14 @@ import type { Installer } from "./index";
 import path from "path";
 import fs from "fs-extra";
 import { PKG_ROOT } from "../consts";
-import { installPkgs } from "../helpers/installPackages";
+import { runPkgManagerInstall } from "../utils/runPkgManagerInstall";
 
 export const trpcInstaller: Installer = async (
   projectDir,
   packageManager,
   packages,
 ) => {
-  await installPkgs({
+  await runPkgManagerInstall({
     packageManager,
     projectDir,
     packages: [
